Refactor SignupForm to make form markup reusable
